+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # 2.0.0-rc.0
 * Log a warning when no `reflector` is provided.
+* Add `AngelEnvironment` class.
+    * Add `Angel.environment`.
+    * Deprecated `app.isProduction` in favor of `app.environment.isProduction`.
 
 # 2.0.0-alpha.24
 * Add `AngelEnv` class to `core`.